# Improve the profiler.