--- a/korangar/src/interface/windows/debug/inspector.rs
+++ b/korangar/src/interface/windows/debug/inspector.rs
@@ -1,4 +1,4 @@
-use korangar_debug::profiling::Measurement;
+use korangar_debug::profiling::FrameMeasurement;
 use korangar_interface::elements::ElementWrap;
 use korangar_interface::size_bound;
 use korangar_interface::windows::{PrototypeWindow, Window, WindowBuilder};
@@ -9,12 +9,12 @@ use crate::interface::layout::ScreenSize;
 use crate::interface::windows::WindowCache;
 
 pub struct FrameInspectorWindow {
-    measurement: Measurement,
+    frame_measurement: FrameMeasurement,
 }
 
 impl FrameInspectorWindow {
-    pub fn new(measurement: Measurement) -> Self {
-        Self { measurement }
+    pub fn new(frame_measurement: FrameMeasurement) -> Self {
+        Self { frame_measurement }
     }
 }
 
@@ -25,7 +25,7 @@ impl PrototypeWindow<InterfaceSettings> for FrameInspectorWindow {
         application: &InterfaceSettings,
         available_space: ScreenSize,
     ) -> Window<InterfaceSettings> {
-        let elements = vec![FrameInspectorView::new(self.measurement.clone()).wrap()];
+        let elements = vec![FrameInspectorView::new(self.frame_measurement.clone()).wrap()];
 
         WindowBuilder::new()
             .with_title("Frame Inspector".to_string())