--- a/korangar_debug/src/lib.rs
+++ b/korangar_debug/src/lib.rs
@@ -1,5 +1,6 @@
 #![feature(decl_macro)]
 #![feature(let_chains)]
+#![feature(slice_take)]
 #![feature(thread_local)]
 
 #[macro_use]