--- a/korangar_debug/src/profiling/mod.rs
+++ b/korangar_debug/src/profiling/mod.rs
@@ -1,8 +1,10 @@
+mod frame_measurement;
 mod measurement;
 mod profiler;
 mod ring_buffer;
 mod statistics;
 
+pub use self::frame_measurement::FrameMeasurement;
 pub use self::measurement::{ActiveMeasurement, Measurement};
 pub use self::profiler::{LockThreadProfiler, Profiler};
 pub use self::ring_buffer::RingBuffer;